refactor: split UI device renderers into modular components
