+++ src/ui/mod.rs
@@ -16,12 +16,12 @@ pub mod buffer;
 pub mod chrome;
 pub mod constants;
 pub mod dashboard;
-pub mod device_renderers;
 pub mod help;
 pub mod layout;
 pub mod notification;
 pub mod process_renderer;
 pub mod renderer;
+pub mod renderers;
 pub mod tabs;
 pub mod text;
 pub mod widgets;
